requirements.txt
absl-py==0.4.0
astor==0.7.1
certifi==2018.11.29
cffi==1.12.1
click==8.0.0
cmake==3.12.0
colorama==0.4.4
common-cmplr-lib-rt==2021.2.0
common-cmplr-lic-rt==2021.2.0
dlib==19.15.0
dpcpp-cpp-rt==2021.2.0
face-recognition==1.2.3
face-recognition-models==0.3.0
ffmpy==0.2.2
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
intel-openmp==2021.2.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
Markdown==2.6.11
mkl==2021.2.0
mkl-fft==1.3.0
mkl-random==1.2.1
munch==2.3.2
numpy==1.16.2
nvidia-ml-py3==7.352.0
olefile==0.46
opencl-rt==2021.2.0
opencv-python==3.4.2.16
pathlib==1.0.1
Pillow==5.4.1
pretrainedmodels==0.7.4
protobuf==3.6.1
pycparser==2.19
PyYAML==5.4.1
scandir==1.7
scipy==1.5.4
six==1.12.0
tbb==2021.2.0
termcolor==1.1.0
torch==1.8.1+cu102
torch-summary==1.4.5
torchvision==0.9.1+cu102
tqdm==4.25.0
typing-extensions==3.10.0.0
wincertstore==0.2
